Fix unexpected behaviour when running tests from a directory.
When running tests by clicking on the directory with subdirectories, the default assumption should be all tests inside that directory + all subdirectories should be executed. However, the reality is those within the subdirectories are skipped. This unexpected behaviour caused us problems when we didn't have a TAP project and our submitted code broke some code without us knowing (we did test it locally but nested tests were skipped).

This CL changes this behaviour to running all tests in all subdirectories when clicking into the root directory, and skipping tests from subdirectories when clicking into the BUILD file, which is compatible with the behaviour of Partial Sync.

TESTED=n/a
PiperOrigin-RevId: 363051078